Formats clade-specific simulation output into standard DAISIE list output with complete STT table

Formats clade-specific simulation output into standard DAISIE list output with complete STT table

Usage

DAISIE_format_IW_full_stt(island_replicates, total_time, M, verbose)

Arguments

island_replicates

List output from DAISIE_sim_core_cr(), DAISIE_sim_core_time_dep(), DAISIE_sim_core_cr_shift() or DAISIE_sim_min_type2() functions. Minimally, this must be a list that has as many elements as replicates. Each element must be a list with the elements island_age, not_present and stt_all. stt_all must be a data frame with the column names Time, nI, nA, nC and present.

total_time

Numeric defining the length of the simulation in time units.

M

Numeric defining the size of mainland pool, i.e. the number of species that can potentially colonize the island.

verbose

A numeric vector of length 1, which in simulations and `DAISIEdataprep()` can be `1` or `0`, where `1` gives intermediate output should be printed. For ML functions a numeric determining if intermediate output should be printed. The default: `0` does not print, `1` prints the initial likelihood and the settings that were selected (which parameters are to be optimised, fixed or shifted), `2` prints the same as `1 and also the intermediate output of the parameters and loglikelihood, while `3` the same as `2` and prints intermediate progress during likelihood computation.

Value

List with IW DAISIE simulation output