Colonization and branching times of 5 Eleutherodactylus (frogs) clades from the island of Hispaniola.

A table containing the colonization and branching times of the Eleutherodacytlus frogs of the island of Hispaniola (Greater Antilles). Each row on the table represents and independent colonisation event. The table has four columns.

$Clade_name - name of independent colonization event
$Status - One of the following categories:
* Non_endemic: for non-endemic island species when an approximate time of colonisation is known
* Non_endemic_MaxAge: for non-endemic island species when colonisation time is unknown
* Endemic: for endemic species when an approximate colonisation time is known
* "Endemic_MaxAge": applies to endemic species or endemic clades for cases where the colonisation time is unknown, or when the user wants to specify an upper bound for colonisation. This could for example apply to endemic species that have recently gone extinct because of anthropogenic causes, and which are not included in the phylogeny ("NA" should be given in the branching times column). It could also apply to insular radiations with long stem branches, for which the time of the first cladogenetic event is known, but the precise time of colonisation is not.
* Endemic&Non_Endemic: when endemic clade and mainland ancestor has re-colonized

Format

A table with 5 rows and 4 columns.

Source

Etienne RS, Haegeman B, Dugo-Cota A, Vila C, Gonzalez-Voyer A & Valente L. The limits to ecological limits to diversification.