Calculates the maximum rates for a Gillespie simulation

Internal function that updates the all the max rates at time t.

Usage

update_max_rates(
  gam,
  laa,
  lac,
  mu,
  hyper_pars = NULL,
  extcutoff,
  K,
  num_spec,
  num_immigrants,
  mainland_n,
  Amin,
  Amax
)

Arguments

gam

A numeric with the per capita immigration rate.

laa

A numeric with the per capita anagenesis rate.

lac

A numeric with the per capita cladogenesis rate.

mu

A numeric with the per capita extinction rate.

hyper_pars

A named list of numeric hyperparameters for the rate calculations as returned by create_hyper_pars():

[1]: is d the scaling parameter for exponent for calculating cladogenesis rate

[2]: is x the exponent for calculating extinction rate

extcutoff

A numeric with the cutoff for the the maximum extinction rate preventing it from being too large and slowing down simulation.

K

A numeric with carrying capacity.

num_spec

A numeric with the current number of species.

num_immigrants

A numeric with the current number of non-endemic species (a.k.a non-endemic species).

mainland_n

A numeric stating the number of mainland species, that is the number of species that can potentially colonize the island. If using a clade-specific diversity dependence, this value is set to 1. If using an island-wide diversity dependence, this value is set to the number of mainland species.

Amin

A numeric value for minimum island area during the simulation.

Amax

A numeric value for maximum island area during the simulation.

Value

a named list with the updated effective rates.

See also