DAISIE: Dynamical Assembly of Islands by Speciation, Immigration and Extinction

DAISIE is an island biogeography model that assumes island biota assembly is governed by immigration, extinction and speciation through cladogenesis and anagenesis. This likelihood-based statistical package can simulate islands and estimate parameters of the DAISIE model based on phylogenetic/phylogeographic data. Cladogenesis and immigration rates can be dependent on diversity.

References

Valente, L. M., Etienne, R. S., & Phillimore, A. B. (2014). The effects of island ontogeny on species diversity and phylogeny. Proceedings of the Royal Society of London. Series B, Biological Sciences 281, 20133227. doi:10.1098/rspb.2013.3227 .

Valente, L., A.B. Phillimore & R.S. Etienne (2015). Equilibrium and non-equilibrium dynamics simultaneously operate in the Galapagos islands. Ecology Letters 18: 844-852, doi:10.1111/ele.12461 .

Valente, L., R.S. Etienne & L. Davalos (2017). Recent extinctions disturb path to equilibrium diversity in Caribbean bats. Nature Ecology and Evolution 1: 0026. doi:10.1038/s41559-016-0026 .

Valente, L., Illera, J. C., Havenstein, K., Pallien, T., Etienne, R. S., & Tiedemann, R. (2017). Equilibrium Bird Species Diversity in Atlantic Islands. Current Biology 27: 1660-1666. doi:10.1016/j.cub.2017.04.053 .

Valente, L., Phillimore, A. B., & Etienne, R. (2018). Using molecular phylogenies in island biogeography: It's about time. Ecography 41: 1684-1686. doi:10.1111/ecog.03503 .

Valente, L., Etienne, R. S., & Garcia-R, J. C. (2019). Deep macroevolutionary impact of humans on New Zealand's unique avifauna. Current Biology 29: 2563-2569. doi:10.1016/j.cub.2019.06.058 .

Valente, L., Phillimore, A. B., Melo, M., Warren, B. H., Clegg, S. M., Havenstein, K., & Etienne, R. S. (2020). A simple dynamic model explains the diversity of island birds worldwide. Nature 579: 92-96. doi:10.1038/s41586-020-2022-5 .

Hauffe, T., Delicado, D., Etienne, R.S., & Valente, L. (2020). Lake expansion elevates equilibrium diversity via increasing colonization. Journal of Biogeography 47: 1849–1860. doi:10.1111/jbi.13914 .

Valente, L., Kristensen, N., Phillimore, A. B., & Etienne, R. S. (2021). Report of programming bugs in the DAISIE R package: consequences and correction. EcoEvoRxiv. doi:10.32942/osf.io/w5ntf .

Santos Neves, P., Lambert, J. W., Valente, L., & Etienne, R. S. (2022). The robustness of a simple dynamic model of island biodiversity to geological and sea-level change. Journal of Biogeography 49: 2091-2104. doi:10.1111/jbi.14519 .

Lambert, J. W., Santos Neves, P., Bilderbeek, R. L. C., Valente, L., Etienne, R. S. (2022). The effect of mainland dynamics on data and parameter estimates in island biogeography. bioRxiv. doi:10.1101/2022.01.13.476210 .

Xie, S., Valente, L., Etienne, R. S. (2023). Can we ignore trait-dependent colonization and diversification in island biogeography? Evolution. doi:10.1093/evolut/qpad006 .

Michielsen, N. M., Goodman, S. M., Soarimalala, V., van der Geer, A. A. E., Dávalos, L. M., Saville, G. I., Upham, N., Valente, L. (2023). The macroevolutionary impact of recent and imminent mammal extinctions on Madagascar. Nature Communications 14 (14).doi:10.1038/s41467-022-35215-3 .

See also

Author

Maintainer: Rampal S. Etienne r.s.etienne@rug.nl (ORCID)

Authors:

Luis Valente luis.valente@naturalis.nl (ORCID)

Albert B. Phillimore (ORCID)

Bart Haegeman (ORCID)

Joshua W. Lambert j.w.l.lambert@rug.nl (ORCID)

Pedro Santos Neves p.m.santos.neves@rug.nl (ORCID)

Shu Xie s.xie@rug.nl (ORCID)

Richèl J.C. Bilderbeek richel@richelbilderbeek.nl (ORCID)

Hanno Hildenbrandt h.hildenbrandt@rug.nl (ORCID)

Other contributors:

Torsten Hauffe torsten.hauffe@gmail.com (ORCID) [contributor]

Giovanni Laudanno glaudanno@gmail.com (ORCID) [contributor]

Nadiah Kristensen nadiah@nadiah.org (ORCID) [contributor]

Ornela Dehayem Nanwou o.d.nanwou@rug.nl (ORCID) [contributor]

Raphael Scherrer r.scherrer@rug.nl (ORCID) [contributor]